+++ ocrd/ocrd/workspace.py
@@ -3,7 +3,7 @@ from os import makedirs, unlink
 from pathlib import Path
 
 import cv2
-from PIL import Image, ImageStat
+from PIL import Image
 import numpy as np
 from atomicwrites import atomic_write
 from deprecated.sphinx import deprecated
@@ -13,7 +13,6 @@ from ocrd_utils import (
     getLogger,
     image_from_polygon,
     coordinates_of_segment,
-    transform_coordinates,
     adjust_canvas_to_rotation,
     adjust_canvas_to_transposition,
     shift_coordinates,
@@ -25,7 +24,6 @@ from ocrd_utils import (
     bbox_from_polygon,
     polygon_from_points,
     xywh_from_bbox,
-    xywh_from_points,
     pushd_popd,
     MIME_TO_EXT,
 )
@@ -59,11 +57,6 @@ class Workspace():
         self.automatic_backup = automatic_backup
         self.baseurl = baseurl
         #  print(mets.to_xml(xmllint=True).decode('utf-8'))
-        self.image_cache = {
-            'pil': {},
-            'cv2': {},
-            'exif': {},
-        }
 
     def __str__(self):
         return 'Workspace[directory=%s, baseurl=%s, file_groups=%s, files=%s]' % (
@@ -103,19 +96,23 @@ class Workspace():
         """
         log.debug('download_file %s [_recursion_count=%s]' % (f, _recursion_count))
         with pushd_popd(self.directory):
-            # XXX FIXME hacky
-            basename = '%s%s' % (f.ID, MIME_TO_EXT.get(f.mimetype, '')) if f.ID else f.basename
             try:
-                f.url = self.resolver.download_to_directory(self.directory, f.url, subdir=f.fileGrp, basename=basename)
-            except FileNotFoundError as e:
-                if not self.baseurl:
-                    raise Exception("No baseurl defined by workspace. Cannot retrieve '%s'" % f.url)
-                if _recursion_count >= 1:
-                    raise Exception("Already tried prepending baseurl '%s'. Cannot retrieve '%s'" % (self.baseurl, f.url))
-                log.debug("First run of resolver.download_to_directory(%s) failed, try prepending baseurl '%s': %s", f.url, self.baseurl, e)
-                f.url = '%s/%s' % (self.baseurl, f.url)
-                f.url = self.download_file(f, _recursion_count + 1).local_filename
-            # XXX FIXME HACK
+                # If the f.url is already a file path, and is within self.directory, do nothing
+                url_path = Path(f.url).resolve()
+                if not (url_path.exists() and url_path.relative_to(str(Path(self.directory).resolve()))):
+                    raise Exception("Not already downloaded, moving on")
+            except Exception as e:
+                basename = '%s%s' % (f.ID, MIME_TO_EXT.get(f.mimetype, '')) if f.ID else f.basename
+                try:
+                    f.url = self.resolver.download_to_directory(self.directory, f.url, subdir=f.fileGrp, basename=basename)
+                except FileNotFoundError as e:
+                    if not self.baseurl:
+                        raise Exception("No baseurl defined by workspace. Cannot retrieve '%s'" % f.url)
+                    if _recursion_count >= 1:
+                        raise Exception("Already tried prepending baseurl '%s'. Cannot retrieve '%s'" % (self.baseurl, f.url))
+                    log.debug("First run of resolver.download_to_directory(%s) failed, try prepending baseurl '%s': %s", f.url, self.baseurl, e)
+                    f.url = '%s/%s' % (self.baseurl, f.url)
+                    f.url = self.download_file(f, _recursion_count + 1).local_filename
             f.local_filename = f.url
             return f
 
@@ -172,8 +169,9 @@ class Workspace():
 
         with pushd_popd(self.directory):
             if 'local_filename' in kwargs:
+                # If the local filename has folder components, create those folders
                 local_filename_dir = kwargs['local_filename'].rsplit('/', 1)[0]
-                if not Path(local_filename_dir).is_dir():
+                if local_filename_dir != kwargs['local_filename'] and not Path(local_filename_dir).is_dir():
                     makedirs(local_filename_dir)
                 if 'url' not in kwargs:
                     kwargs['url'] = kwargs['local_filename']
@@ -212,11 +210,9 @@ class Workspace():
         files = self.mets.find_files(url=image_url)
         f = files[0] if files else OcrdFile(None, url=image_url)
         image_filename = self.download_file(f).local_filename
-
-        if image_url not in self.image_cache['exif']:
-            # FIXME must be in the right directory
-            self.image_cache['exif'][image_url] = OcrdExif(Image.open(image_filename))
-        return self.image_cache['exif'][image_url]
+        with Image.open(image_filename) as pil_img:
+            ocrd_exif = OcrdExif(pil_img)
+        return ocrd_exif
 
     @deprecated(version='1.0.0', reason="Use workspace.image_from_page and workspace.image_from_segment")
     def resolve_image_as_pil(self, image_url, coords=None):
@@ -237,20 +233,17 @@ class Workspace():
         f = files[0] if files else OcrdFile(None, url=image_url)
         image_filename = self.download_file(f).local_filename
 
-        if image_url not in self.image_cache['pil']:
-            with pushd_popd(self.directory):
-                self.image_cache['pil'][image_url] = Image.open(image_filename)
-
-        pil_image = self.image_cache['pil'][image_url]
+        with pushd_popd(self.directory):
+            pil_image = Image.open(image_filename)
 
         if coords is None:
             return pil_image
-        if image_url not in self.image_cache['cv2']:
-            log.debug("Converting PIL to OpenCV: %s", image_url)
-            color_conversion = cv2.COLOR_GRAY2BGR if pil_image.mode in ('1', 'L') else  cv2.COLOR_RGB2BGR
-            pil_as_np_array = np.array(pil_image).astype('uint8') if pil_image.mode == '1' else np.array(pil_image)
-            self.image_cache['cv2'][image_url] = cv2.cvtColor(pil_as_np_array, color_conversion)
-        cv2_image = self.image_cache['cv2'][image_url]
+
+        log.debug("Converting PIL to OpenCV: %s", image_url)
+        color_conversion = cv2.COLOR_GRAY2BGR if pil_image.mode in ('1', 'L') else  cv2.COLOR_RGB2BGR
+        pil_as_np_array = np.array(pil_image).astype('uint8') if pil_image.mode == '1' else np.array(pil_image)
+        cv2_image = cv2.cvtColor(pil_as_np_array, color_conversion)
+
         poly = np.array(coords, np.int32)
         log.debug("Cutting region %s from %s", coords, image_url)
         region_cut = cv2_image[