+++ ocrd_utils/ocrd_utils/constants.py
@@ -18,9 +18,17 @@ EXT_TO_MIME = {
     '.tif': 'image/tiff',
     '.tiff': 'image/tiff',
     '.png': 'image/png',
-    '.jpg': 'image/jpg',
-    '.jpeg': 'image/jpg',
-    '.xml': MIMETYPE_PAGE
+    '.jpg': 'image/jpeg',
+    '.jpeg': 'image/jpeg',
+    '.xml': MIMETYPE_PAGE,
+    '.jp2': 'image/jp2',
+    '.pdf': 'application/pdf',
+    '.ps': 'application/postscript',
+    '.eps': 'application/postscript',
+    '.xps': 'application/oxps',
+    '.ppm': 'image/x-portable-pixmap',
+    '.pnm': 'image/x-portable-anymap',
+    '.pbm': 'image/x-portable-bitmap',
 }
 
 MIME_TO_EXT = {
@@ -30,4 +38,11 @@ MIME_TO_EXT = {
     'image/jpeg': '.jpg',
     MIMETYPE_PAGE: '.xml',
     'application/alto+xml': '.xml',
+    'image/jp2': '.jp2',
+    'application/pdf': '.pdf',
+    'application/postscript': '.ps',
+    'application/oxps': '.xps',
+    'image/x-portable-pixmap': '.ppm',
+    'image/x-portable-anymap': '.pnm',
+    'image/x-portable-bitmap': '.pbm',
 }