+++ ocrd_validators/ocrd_validators/workspace_validator.py
@@ -4,6 +4,7 @@ Validating a workspace.
 import re
 from traceback import format_exc
 from pathlib import Path
+from contextlib import contextmanager
 
 from ocrd_utils import getLogger, MIMETYPE_PAGE, pushd_popd, is_local_filename
 from ocrd_modelfactory import page_from_file
@@ -22,6 +23,31 @@ class WorkspaceValidator():
     """
     Validates an OCR-D/METS workspace against the specs.
     """
+
+    @staticmethod
+    def check_file_grp(workspace, input_file_grp=None, output_file_grp=None, report=None):
+        """
+        Return a report on whether input_file_grp is/are in workspace.mets and output_file_grp is/are not.
+        To be run before processing
+        """
+        if not report:
+            report = ValidationReport()
+        if isinstance(input_file_grp, str):
+            input_file_grp = input_file_grp.split(',')
+        if isinstance(output_file_grp, str):
+            output_file_grp = output_file_grp.split(',')
+
+        log.info("input_file_grp=%s output_file_grp=%s" % (input_file_grp, output_file_grp))
+        print(input_file_grp)
+        if input_file_grp:
+            for grp in input_file_grp:
+                if grp not in workspace.mets.file_groups:
+                    report.add_error("Input fileGrp[@USE='%s'] not in METS!" % grp)
+        if output_file_grp:
+            for grp in output_file_grp:
+                if grp in workspace.mets.file_groups:
+                    report.add_error("Output fileGrp[@USE='%s'] already in METS!" % grp)
+        return report
 
     def __init__(self, resolver, mets_url, src_dir=None, skip=None, download=False,
                  page_strictness='strict', page_coordinate_consistency='poly'):
@@ -200,9 +226,9 @@ class WorkspaceValidator():
                 if '-' in category:
                     category, name = category.split('-', 1)
                 if category not in FILE_GROUP_CATEGORIES:
-                    self.report.add_error("Unspecified USE category '%s' in fileGrp '%s'" % (category, fileGrp))
+                    self.report.add_warning("Unspecified USE category '%s' in fileGrp '%s'" % (category, fileGrp))
                 if name is not None and not re.match(r'^[A-Z0-9-]{3,}$', name):
-                    self.report.add_error("Invalid USE name '%s' in fileGrp '%s'" % (name, fileGrp))
+                    self.report.add_warning("Invalid USE name '%s' in fileGrp '%s'" % (name, fileGrp))
 
     def _validate_mets_files(self):
         """